+++ Storage/Storage/Model/StatsVersionBySite.swift
@@ -2,9 +2,9 @@
 /// These entities will be serialised to a plist file
 ///
 public struct StatsVersionBySite: Codable, Equatable {
-    public let statsVersionBySite: [Int: StatsVersion]
+    public let statsVersionBySite: [Int64: StatsVersion]
 
-    public init(statsVersionBySite: [Int: StatsVersion]) {
+    public init(statsVersionBySite: [Int64: StatsVersion]) {
         self.statsVersionBySite = statsVersionBySite
     }
 }
