+++ Yosemite/Yosemite/Stores/OrderNoteStore.swift
@@ -42,7 +42,7 @@ private extension OrderNoteStore {
 
     /// Retrieves the order notes associated with the provided Site ID & Order ID (if any!).
     ///
-    func retrieveOrderNotes(siteID: Int, orderID: Int, onCompletion: @escaping ([OrderNote]?, Error?) -> Void) {
+    func retrieveOrderNotes(siteID: Int64, orderID: Int64, onCompletion: @escaping ([OrderNote]?, Error?) -> Void) {
         let remote = OrdersRemote(network: network)
         remote.loadOrderNotes(for: siteID, orderID: orderID) { [weak self] (orderNotes, error) in
             guard let orderNotes = orderNotes else {
@@ -58,7 +58,7 @@ private extension OrderNoteStore {
 
     /// Adds a single order note and associates it with the provided siteID and orderID.
     ///
-    func addOrderNote(siteID: Int, orderID: Int, isCustomerNote: Bool, note: String, onCompletion: @escaping (OrderNote?, Error?) -> Void) {
+    func addOrderNote(siteID: Int64, orderID: Int64, isCustomerNote: Bool, note: String, onCompletion: @escaping (OrderNote?, Error?) -> Void) {
         let remote = OrdersRemote(network: network)
         remote.addOrderNote(for: siteID, orderID: orderID, isCustomerNote: isCustomerNote, with: note) { [weak self] (orderNote, error) in
             guard let note = orderNote else {
@@ -80,7 +80,7 @@ extension OrderNoteStore {
 
     /// Updates (OR Inserts) the specified ReadOnly OrderNote Entity into the Storage Layer.
     ///
-    func upsertStoredOrderNoteInBackground(readOnlyOrderNote: Networking.OrderNote, orderID: Int, onCompletion: @escaping () -> Void) {
+    func upsertStoredOrderNoteInBackground(readOnlyOrderNote: Networking.OrderNote, orderID: Int64, onCompletion: @escaping () -> Void) {
         let derivedStorage = sharedDerivedStorage
         derivedStorage.perform {
             self.saveNote(derivedStorage, readOnlyOrderNote, orderID)
@@ -93,7 +93,7 @@ extension OrderNoteStore {
 
     /// Updates (OR Inserts) the specified ReadOnly OrderNote Entities into the Storage Layer.
     ///
-    func upsertStoredOrderNotesInBackground(readOnlyOrderNotes: [Networking.OrderNote], orderID: Int, onCompletion: @escaping () -> Void) {
+    func upsertStoredOrderNotesInBackground(readOnlyOrderNotes: [Networking.OrderNote], orderID: Int64, onCompletion: @escaping () -> Void) {
         let derivedStorage = sharedDerivedStorage
         derivedStorage.perform {
             for readOnlyOrderNote in readOnlyOrderNotes {
@@ -109,7 +109,7 @@ extension OrderNoteStore {
     /// Using the provided StorageType, update or insert a Storage.OrderNote using the provided ReadOnly
     /// OrderNote. This func does *not* persist any unsaved changes to storage.
     ///
-    private func saveNote(_ storage: StorageType, _ readOnlyOrderNote: OrderNote, _ orderID: Int) {
+    private func saveNote(_ storage: StorageType, _ readOnlyOrderNote: OrderNote, _ orderID: Int64) {
         if let existingStorageNote = storage.loadOrderNote(noteID: readOnlyOrderNote.noteID) {
             existingStorageNote.update(with: readOnlyOrderNote)
             return
